Implement bootable engine core with Q2-based subsystems
Working engine that builds with MSVC + SDL2 and boots to main loop:
- Zone memory allocator (Z_Malloc/Z_Free/Z_Touch, exported API)
- Command buffer and console command system (Cmd_AddCommand, exec, echo)
- Console variable system (Cvar_Get/Set/SetValue, archive, latch, userinfo)
- PAK file virtual filesystem (FS_FOpenFile/LoadFile/FreeFile)
- Main entry point replacing WinMain with SDL2 main loop
- Engine init sequence: Z_Init → Cbuf → Cmd → Cvar → FS → config exec
- Shared utilities: vector math, string ops, info strings, byte order, RNG

Builds to 27KB sof.exe that successfully initializes and enters the frame
loop. All 24 SoF.exe exports (consumed by sound/game DLLs) are implemented
as direct function calls in the unified binary.

+++ src/engine/sys_sdl.c
@@ -76,9 +76,9 @@ void Sys_PlatformShutdown(void)
    Timer
    ========================================================================== */
 
-uint32_t Sys_Milliseconds(void)
+int Sys_Milliseconds(void)
 {
-    return SDL_GetTicks();
+    return (int)SDL_GetTicks();
 }
 
 /* ==========================================================================
